Extract out common code form standalone
