+++ pulsar-broker/src/main/java/org/apache/pulsar/PulsarStandaloneStarter.java
@@ -21,83 +21,16 @@ package org.apache.pulsar;
 import static org.apache.commons.lang3.StringUtils.isBlank;
 
 import java.io.FileInputStream;
-import java.net.URL;
 
-import java.nio.file.Paths;
-import java.util.Optional;
-import org.apache.pulsar.broker.PulsarService;
 import org.apache.pulsar.broker.ServiceConfiguration;
 import org.apache.pulsar.broker.ServiceConfigurationUtils;
-import org.apache.pulsar.client.admin.PulsarAdmin;
-import org.apache.pulsar.client.admin.PulsarAdminException;
 import org.apache.pulsar.common.configuration.PulsarConfigurationLoader;
-import org.apache.pulsar.common.naming.TopicName;
-import org.apache.pulsar.common.policies.data.ClusterData;
-import org.apache.pulsar.common.policies.data.TenantInfo;
-import org.apache.pulsar.functions.worker.WorkerConfig;
-import org.apache.pulsar.functions.worker.WorkerService;
-import org.apache.pulsar.zookeeper.LocalBookkeeperEnsemble;
-import org.aspectj.weaver.loadtime.Agent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.beust.jcommander.JCommander;
-import com.beust.jcommander.Parameter;
-import com.ea.agentloader.AgentLoader;
-import com.google.common.collect.Sets;
 
-public class PulsarStandaloneStarter {
-
-    PulsarService broker;
-    PulsarAdmin admin;
-    LocalBookkeeperEnsemble bkEnsemble;
-    ServiceConfiguration config;
-    WorkerService fnWorkerService;
-
-    @Parameter(names = { "-c", "--config" }, description = "Configuration file path", required = true)
-    private String configFile;
-
-    @Parameter(names = { "--wipe-data" }, description = "Clean up previous ZK/BK data")
-    private boolean wipeData = false;
-
-    @Parameter(names = { "--num-bookies" }, description = "Number of local Bookies")
-    private int numOfBk = 1;
-
-    @Parameter(names = { "--zookeeper-port" }, description = "Local zookeeper's port")
-    private int zkPort = 2181;
-
-    @Parameter(names = { "--bookkeeper-port" }, description = "Local bookies base port")
-    private int bkPort = 3181;
-
-    @Parameter(names = { "--zookeeper-dir" }, description = "Local zooKeeper's data directory")
-    private String zkDir = "data/standalone/zookeeper";
-
-    @Parameter(names = { "--bookkeeper-dir" }, description = "Local bookies base data directory")
-    private String bkDir = "data/standalone/bookkeeper";
-
-    @Parameter(names = { "--no-broker" }, description = "Only start ZK and BK services, no broker")
-    private boolean noBroker = false;
-
-    @Parameter(names = { "--only-broker" }, description = "Only start Pulsar broker service (no ZK, BK)")
-    private boolean onlyBroker = false;
-
-    @Parameter(names = {"-nfw", "--no-functions-worker"}, description = "Run functions worker with Broker")
-    private boolean noFunctionsWorker = false;
-
-    @Parameter(names = {"-fwc", "--functions-worker-conf"}, description = "Configuration file for Functions Worker")
-    private String fnWorkerConfigFile = Paths.get("").toAbsolutePath().normalize().toString() + "/conf/functions_worker.yml";
-
-    @Parameter(names = {"-nss", "--no-stream-storage"}, description = "Disable stream storage")
-    private boolean noStreamStorage = false;
-
-    @Parameter(names = { "--stream-storage-port" }, description = "Local bookies stream storage port")
-    private int streamStoragePort = 4181;
-
-    @Parameter(names = { "-a", "--advertised-address" }, description = "Standalone broker advertised address")
-    private String advertisedAddress = null;
-
-    @Parameter(names = { "-h", "--help" }, description = "Show this help message")
-    private boolean help = false;
+public class PulsarStandaloneStarter extends StandalonePulsarBase {
 
     private static final Logger log = LoggerFactory.getLogger(PulsarStandaloneStarter.class);
 
@@ -107,12 +40,12 @@ public class PulsarStandaloneStarter {
         try {
             jcommander.addObject(this);
             jcommander.parse(args);
-            if (help || isBlank(configFile)) {
+            if (this.isHelp() || isBlank(this.getConfigFile())) {
                 jcommander.usage();
                 return;
             }
 
-            if (noBroker && onlyBroker) {
+            if (this.isNoBroker() && this.isOnlyBroker()) {
                 log.error("Only one option is allowed between '--no-broker' and '--only-broker'");
                 jcommander.usage();
                 return;
@@ -122,14 +55,14 @@ public class PulsarStandaloneStarter {
             return;
         }
 
-        this.config = PulsarConfigurationLoader.create((new FileInputStream(configFile)), ServiceConfiguration.class);
+        this.config = PulsarConfigurationLoader.create((new FileInputStream(this.getConfigFile())), ServiceConfiguration.class);
 
         String zkServers = "127.0.0.1";
 
-        if (advertisedAddress != null) {
+        if (this.getAdvertisedAddress() != null) {
             // Use advertised address from command line
-            config.setAdvertisedAddress(advertisedAddress);
-            zkServers = advertisedAddress;
+            config.setAdvertisedAddress(this.getAdvertisedAddress());
+            zkServers = this.getAdvertisedAddress();
         } else if (isBlank(config.getAdvertisedAddress())) {
             // Use advertised address as local hostname
             config.setAdvertisedAddress(ServiceConfigurationUtils.unsafeLocalhostResolve());
@@ -138,8 +71,8 @@ public class PulsarStandaloneStarter {
         }
 
         // Set ZK server's host to localhost
-        config.setZookeeperServers(zkServers + ":" + zkPort);
-        config.setConfigurationStoreServers(zkServers + ":" + zkPort);
+        config.setZookeeperServers(zkServers + ":" + this.getZkPort());
+        config.setConfigurationStoreServers(zkServers + ":" + this.getZkPort());
         config.setRunningStandalone(true);
 
         Runtime.getRuntime().addShutdownHook(new Thread() {
@@ -161,111 +94,6 @@ public class PulsarStandaloneStarter {
                 }
             }
         });
-    }
-
-    void start() throws Exception {
-
-        if (config == null) {
-            System.exit(1);
-        }
-
-        log.debug("--- setup PulsarStandaloneStarter ---");
-
-        // load aspectj-weaver agent for instrumentation
-        AgentLoader.loadAgentClass(Agent.class.getName(), null);
-
-        if (!onlyBroker) {
-            // Start LocalBookKeeper
-            bkEnsemble = new LocalBookkeeperEnsemble(
-                numOfBk, zkPort, bkPort, streamStoragePort, zkDir, bkDir, wipeData, config.getAdvertisedAddress());
-            bkEnsemble.startStandalone(!noStreamStorage);
-        }
-
-        if (noBroker) {
-            return;
-        }
-
-        // initialize the functions worker
-        if (!noFunctionsWorker) {
-            WorkerConfig workerConfig;
-            if (isBlank(fnWorkerConfigFile)) {
-                workerConfig = new WorkerConfig();
-            } else {
-                workerConfig = WorkerConfig.load(fnWorkerConfigFile);
-            }
-            // worker talks to local broker
-            workerConfig.setPulsarServiceUrl("pulsar://127.0.0.1:" + config.getBrokerServicePort());
-            workerConfig.setPulsarWebServiceUrl("http://127.0.0.1:" + config.getWebServicePort());
-            workerConfig.setStateStorageServiceUrl("bk://127.0.0.1:" + streamStoragePort);
-            String hostname = ServiceConfigurationUtils.getDefaultOrConfiguredAddress(
-                config.getAdvertisedAddress());
-            workerConfig.setWorkerHostname(hostname);
-            workerConfig.setWorkerId(
-                "c-" + config.getClusterName()
-                    + "-fw-" + hostname
-                    + "-" + workerConfig.getWorkerPort());
-            fnWorkerService = new WorkerService(workerConfig);
-        }
-
-        // Start Broker
-        broker = new PulsarService(config, Optional.ofNullable(fnWorkerService));
-        broker.start();
-
-        URL webServiceUrl = new URL(
-                String.format("http://%s:%d", config.getAdvertisedAddress(), config.getWebServicePort()));
-        final String brokerServiceUrl = String.format("pulsar://%s:%d", config.getAdvertisedAddress(),
-                config.getBrokerServicePort());
-        admin = PulsarAdmin.builder().serviceHttpUrl(webServiceUrl.toString()).authentication(
-                config.getBrokerClientAuthenticationPlugin(), config.getBrokerClientAuthenticationParameters()).build();
-
-        // Create a sample namespace
-        final String property = "sample";
-        final String cluster = config.getClusterName();
-        final String globalCluster = "global";
-        final String namespace = property + "/" + cluster + "/ns1";
-        try {
-            ClusterData clusterData = new ClusterData(webServiceUrl.toString(), null /* serviceUrlTls */,
-                    brokerServiceUrl, null /* brokerServiceUrlTls */);
-            if (!admin.clusters().getClusters().contains(cluster)) {
-                admin.clusters().createCluster(cluster, clusterData);
-            } else {
-                admin.clusters().updateCluster(cluster, clusterData);
-            }
-
-            // Create marker for "global" cluster
-            if (!admin.clusters().getClusters().contains(globalCluster)) {
-                admin.clusters().createCluster(globalCluster, new ClusterData(null, null));
-            }
-
-            if (!admin.tenants().getTenants().contains(property)) {
-                admin.tenants().createTenant(property,
-                        new TenantInfo(Sets.newHashSet(config.getSuperUserRoles()), Sets.newHashSet(cluster)));
-            }
-
-            if (!admin.namespaces().getNamespaces(property).contains(namespace)) {
-                admin.namespaces().createNamespace(namespace);
-            }
-        } catch (PulsarAdminException e) {
-            log.info(e.getMessage());
-        }
-
-        // Create a public tenant and default namespace
-        final String publicTenant = TopicName.PUBLIC_TENANT;
-        final String defaultNamespace = TopicName.PUBLIC_TENANT + "/" + TopicName.DEFAULT_NAMESPACE;
-        try {
-            if (!admin.tenants().getTenants().contains(publicTenant)) {
-                admin.tenants().createTenant(publicTenant,
-                        new TenantInfo(Sets.newHashSet(config.getSuperUserRoles()), Sets.newHashSet(cluster)));
-            }
-            if (!admin.namespaces().getNamespaces(publicTenant).contains(defaultNamespace)) {
-                admin.namespaces().createNamespace(defaultNamespace);
-                admin.namespaces().setNamespaceReplicationClusters(defaultNamespace, Sets.newHashSet(config.getClusterName()));
-            }
-        } catch (PulsarAdminException e) {
-            log.info(e.getMessage());
-        }
-
-        log.debug("--- setup completed ---");
     }
 
     public static void main(String args[]) throws Exception {
